+++ src/main/java/net/minecraft/block/BlockAnvil.java
@@ -59,7 +59,7 @@ public class BlockAnvil extends BlockFalling {
     }
 
     public int damageDropped(IBlockState state) {
-        return state.getValue(DAMAGE).intValue();
+        return state.getValue(DAMAGE);
     }
 
     public void setBlockBoundsBasedOnState(IBlockAccess worldIn, BlockPos pos) {
@@ -101,7 +101,7 @@ public class BlockAnvil extends BlockFalling {
     public int getMetaFromState(IBlockState state) {
         int i = 0;
         i = i | state.getValue(FACING).getHorizontalIndex();
-        i = i | state.getValue(DAMAGE).intValue() << 2;
+        i = i | state.getValue(DAMAGE) << 2;
         return i;
     }
