+++ src/main/java/net/minecraft/block/BlockBanner.java
@@ -191,7 +191,7 @@ public class BlockBanner extends BlockContainer {
         }
 
         public int getMetaFromState(IBlockState state) {
-            return state.getValue(ROTATION).intValue();
+            return state.getValue(ROTATION);
         }
 
         protected BlockState createBlockState() {
