+++ src/main/java/net/minecraft/block/BlockCrops.java
@@ -37,7 +37,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
         super.updateTick(worldIn, pos, state, rand);
 
         if (worldIn.getLightFromNeighbors(pos.up()) >= 9) {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i < 7) {
                 float f = getGrowthChance(this, worldIn, pos);
@@ -50,7 +50,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
     }
 
     public void grow(World worldIn, BlockPos pos, IBlockState state) {
-        int i = state.getValue(AGE).intValue() + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
+        int i = state.getValue(AGE) + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
 
         if (i > 7) {
             i = 7;
@@ -71,7 +71,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
                 if (iblockstate.getBlock() == Blocks.farmland) {
                     f1 = 1.0F;
 
-                    if (iblockstate.getValue(BlockFarmland.MOISTURE).intValue() > 0) {
+                    if (iblockstate.getValue(BlockFarmland.MOISTURE) > 0) {
                         f1 = 3.0F;
                     }
                 }
@@ -120,7 +120,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
         super.dropBlockAsItemWithChance(worldIn, pos, state, chance, 0);
 
         if (!worldIn.isRemote) {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i >= 7) {
                 int j = 3 + fortune;
@@ -135,7 +135,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
     }
 
     public Item getItemDropped(IBlockState state, Random rand, int fortune) {
-        return state.getValue(AGE).intValue() == 7 ? this.getCrop() : this.getSeed();
+        return state.getValue(AGE) == 7 ? this.getCrop() : this.getSeed();
     }
 
     public Item getItem(World worldIn, BlockPos pos) {
@@ -143,7 +143,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
     }
 
     public boolean canGrow(World worldIn, BlockPos pos, IBlockState state, boolean isClient) {
-        return state.getValue(AGE).intValue() < 7;
+        return state.getValue(AGE) < 7;
     }
 
     public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state) {
@@ -159,7 +159,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(AGE).intValue();
+        return state.getValue(AGE);
     }
 
     protected BlockState createBlockState() {
