+++ src/main/java/net/minecraft/block/BlockDaylightDetector.java
@@ -42,7 +42,7 @@ public class BlockDaylightDetector extends BlockContainer {
     }
 
     public int getWeakPower(IBlockAccess worldIn, BlockPos pos, IBlockState state, EnumFacing side) {
-        return state.getValue(POWER).intValue();
+        return state.getValue(POWER);
     }
 
     public void updatePower(World worldIn, BlockPos pos) {
@@ -59,7 +59,7 @@ public class BlockDaylightDetector extends BlockContainer {
                 i = 15 - i;
             }
 
-            if (iblockstate.getValue(POWER).intValue() != i) {
+            if (iblockstate.getValue(POWER) != i) {
                 worldIn.setBlockState(pos, iblockstate.withProperty(POWER, i), 3);
             }
         }
@@ -118,7 +118,7 @@ public class BlockDaylightDetector extends BlockContainer {
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(POWER).intValue();
+        return state.getValue(POWER);
     }
 
     protected BlockState createBlockState() {
