+++ src/main/java/net/minecraft/block/BlockFarmland.java
@@ -42,7 +42,7 @@ public class BlockFarmland extends Block {
     }
 
     public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand) {
-        int i = state.getValue(MOISTURE).intValue();
+        int i = state.getValue(MOISTURE);
 
         if (!this.hasWater(worldIn, pos) && !worldIn.isRainingAt(pos.up())) {
             if (i > 0) {
@@ -122,7 +122,7 @@ public class BlockFarmland extends Block {
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(MOISTURE).intValue();
+        return state.getValue(MOISTURE);
     }
 
     protected BlockState createBlockState() {
