+++ src/main/java/net/minecraft/client/network/NetHandlerPlayClient.java
@@ -1036,7 +1036,7 @@ public class NetHandlerPlayClient implements INetHandlerPlayClient {
 
         for (Entry<StatBase, Integer> entry : packetIn.func_148974_c().entrySet()) {
             StatBase statbase = entry.getKey();
-            int i = entry.getValue().intValue();
+            int i = entry.getValue();
 
             if (statbase.isAchievement() && i > 0) {
                 if (this.field_147308_k && this.gameController.thePlayer.getStatFileWriter().readStat(statbase) == 0) {
