+++ src/main/java/net/minecraft/client/player/inventory/ContainerLocalMenu.java
@@ -22,7 +22,7 @@ public class ContainerLocalMenu extends InventoryBasic implements ILockableConta
     }
 
     public int getField(int id) {
-        return this.field_174895_b.containsKey(id) ? this.field_174895_b.get(id).intValue() : 0;
+        return this.field_174895_b.containsKey(id) ? this.field_174895_b.get(id) : 0;
     }
 
     public void setField(int id, int value) {
