+++ src/main/java/net/minecraft/client/renderer/vertex/VertexFormat.java
@@ -92,7 +92,7 @@ public class VertexFormat {
     }
 
     public int getUvOffsetById(int id) {
-        return this.uvOffsetsById.get(id).intValue();
+        return this.uvOffsetsById.get(id);
     }
 
     public String toString() {
@@ -144,7 +144,7 @@ public class VertexFormat {
     }
 
     public int getOffset(int p_181720_1_) {
-        return this.offsets.get(p_181720_1_).intValue();
+        return this.offsets.get(p_181720_1_);
     }
 
     public boolean equals(Object p_equals_1_) {
